+++ src/adc/filter/__init__.py
@@ -1,5 +1,6 @@
 from ._attach_metadata import AttachMetaData
 from ._base_audio_augmentation import BaseAudioAugmentationFilter, AUG_MODES, AUG_MODE_ADD, AUG_MODE_REPLACE
+from ._change_volume import ChangeVolume
 from ._convert_to_mono import ConvertToMono
 from ._convert_to_wav import ConvertToWav
 from ._discard_negatives import DiscardNegatives